Empty element appears in month and year combobox
When you open the combobox, appears in the beginning an empty element. Appears over January and over 1900. The date that returns when you select this elements is December 1900. In the source code I see that there is an empty field <option> in datepicker.component.html file that it is causing the problem. Can you fix it?