Add Language defaults & Add italian to supported languages
Many translation services uses "de", "it" etc. as language code.
This is useful if you gonna focus on Europe, Asia, etc. because you do
not need to care about other countries.

On ion-datepicker, if you wanna change the language, you always have to
explicitly declare the language, something not every translation service
(see example above) does do, and to write bridges to get the correct
language is annoying for every dev. So default languages are a solution.

Signed-off-by: Tristan Marsell <tristan@sqrls.de>